+++ src/app.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const button = <button onClick={() => { alert('clicked') }}>Add</button>
+
+export default button;